# src/modules/users/dtos/create-user.dto.ts
import { StringField } from "src/decorators/field.decorators";

export class CreateUserDto {
  @StringField()
  readonly email: string;

  @StringField()
  readonly password: string;

  @StringField()
  readonly fullName: string;

  @StringField()
  readonly phoneNumber: string;
}